add ui.image_scaled() + move into image module
